requirements.txt
pinecone-client>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=22.0.0
flask-cors>=4.0.0
pinecone~=5.4.2